# queuefree/services/api/openapi/queuefree.v1.yaml
openapi: 3.0.0
paths:
...
  /v1/queue-guard/check-in:
    post:
      operationId: checkInQueueGuard
      parameters:
        - name: idempotency-key
          required: true
          in: header
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          description: Idempotency key for safely retryable write actions.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserQueueGuardResponseDto'
        '400':
          description: Missing or invalid Idempotency-Key header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Check in and refresh the current user queue guard validity
      tags:
        - QueueGuard
  /v1/orders:
    post:
      operationId: createOrder
      parameters:
        - name: idempotency-key
          required: true
          in: header
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          description: Idempotency key for safely retryable write actions.
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequestDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponseDto'
        '400':
          description: Invalid order creation request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '404':
          description: Product, SKU, or address not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Create a single-item order for the current C-end user
      tags:
        - Orders
  /v1/orders/{orderId}/payment-intents:
    post:
      operationId: createPaymentIntent
      parameters:
        - name: orderId
          required: true
          in: path
          schema:
            example: order_demo_01
            type: string
        - name: idempotency-key
          required: true
          in: header
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          description: Idempotency key for safely retryable write actions.
          required: true
          schema:
            type: string
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePaymentIntentResponseDto'
        '400':
          description: Missing or invalid Idempotency-Key header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '404':
          description: Order not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '409':
          description: Order cannot create a payment intent in its current status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Create a payment intent for an existing order
      tags:
        - Payments
...
info:
  title: QueueFree API
  description: >-
    Initial C-end OpenAPI surface for QueueFree. This batch exposes system-level
    endpoints plus the minimum public rules endpoints needed by Web rules pages
    and packages/api-client.
  version: 0.0.1
  contact: {}
...
servers:
  - url: http://localhost:4000
    description: local
  - url: https://dev-api.queuefree.com
    description: dev
  - url: https://stg-api.queuefree.com
    description: staging
  - url: https://api.queuefree.com
    description: prod
components:
  schemas:
...
    ApiErrorResponseDto:
      type: object
      properties:
        code:
          type: string
          example: INTERNAL_SERVER_ERROR
        message:
          type: string
          example: Unexpected server error.
        requestId:
          type: string
          example: req_demo_123
        details:
          type: object
          additionalProperties: true
          description: >-
            Opaque structured error details for debugging or validation
            feedback.
      required:
        - code
        - message
...
    UserQueueGuardResponseDto:
      type: object
      properties:
        status:
          type: string
        validUntil:
          type: string
        graceUntil:
          type: string
        lastCheckinAt:
          type: object
          nullable: true
        canCheckInNow:
          type: boolean
        activeEntriesCount:
          type: number
        frozenEntriesCount:
          type: number
      required:
        - status
        - validUntil
        - graceUntil
        - lastCheckinAt
        - canCheckInNow
        - activeEntriesCount
        - frozenEntriesCount
    CreateOrderRequestDto:
      type: object
      properties:
        productId:
          type: string
          example: prod_demo_01
        skuId:
          type: string
          example: sku_demo_01
        quantity:
          type: number
          example: 1
        addressId:
          type: string
          example: addr_demo_01
      required:
        - productId
        - skuId
        - quantity
        - addressId
    OrderStatus:
      type: string
      enum:
        - CREATED
        - WAIT_PAY
        - PAID
        - FULFILLING
        - SHIPPED
        - DELIVERED
        - COMPLETED
        - CANCELED
        - AFTERSALE_OPEN
        - PARTIAL_REFUNDED
        - FULL_REFUNDED
    CreateOrderResponseDto:
      type: object
      properties:
        orderId:
          type: string
          example: order_demo_01
        status:
          example: WAIT_PAY
          allOf:
            - $ref: '#/components/schemas/OrderStatus'
        productId:
          type: string
          example: prod_demo_01
        skuId:
          type: string
          example: sku_demo_01
        quantity:
          type: number
          example: 1
      required:
        - orderId
        - status
        - productId
        - skuId
        - quantity
    CreatePaymentIntentResponseDto:
      type: object
      properties:
        paymentIntentId:
          type: string
          example: pi_demo_01
        orderId:
          type: string
          example: order_demo_01
        provider:
          type: string
          example: mockpay
        amountMinor:
          type: number
          example: 129900
          description: Minor unit amount that should be paid for the order.
        currencyCode:
          type: string
          example: PHP
        checkoutUrl:
          type: string
          example: https://checkout.queuefree.com/payment-intents/pi_demo_01
      required:
        - paymentIntentId
        - orderId
        - provider
        - amountMinor
        - currencyCode
        - checkoutUrl
...